stripe payment now posting success, changes to front and back. needs more backend work.
Focus of this commit was to get stripe payments working and see the payments succeed in stripe dashboard. Also reworked flow with frontend and backend so public key is in front end with the payment method and private key in backend with payment intent. There is still work to be done in backend to save the successful payment and recalculate the current balance due.
- http - added new endpoint for payment.
- https-stripe - removed Julian's secret key and added Brantly's public key.
- Account - reworked the stripe api calls, method first then intent, pointed to new endpoint in http file. Added returns for validation functions, because state can't be used properly in the check errors function, also easier to read.
- BalanceController - commented endpoints out and created new /pay endpoint which takes a paymentMethodId and amount, the stripe call in this endpoint processes the payment immediately.